improve the response for big number to string

+++ mall-common/src/main/java/com/oyproj/common/api/CommonResult.java
@@ -9,14 +9,14 @@ import cn.hutool.json.JSONUtil;
  * @github https://github.com/oywq3000
  */
 public class CommonResult<T> {
-    private long code;
+    private int code;
     private String message;
     private T data;
 
     protected CommonResult() {
     }
 
-    protected CommonResult(long code, String message, T data) {
+    protected CommonResult(int code, String message, T data) {
         this.code = code;
         this.message = message;
         this.data = data;
@@ -102,11 +102,11 @@ public class CommonResult<T> {
         return new CommonResult<T>(ResultCode.FORBIDDEN.getCode(), ResultCode.FORBIDDEN.getMessage(), data);
     }
 
-    public long getCode() {
+    public int getCode() {
         return code;
     }
 
-    public void setCode(long code) {
+    public void setCode(int code) {
         this.code = code;
     }
