+++ mall-common/src/main/java/com/oyproj/common/api/IErrorCode.java
@@ -4,7 +4,7 @@ package com.oyproj.common.api;
  * 封装API的错误码
  */
 public interface IErrorCode {
-    long getCode();
+    int getCode();
 
     String getMessage();
 }
